+++ js/upload-form.js
@@ -1,66 +1,69 @@
-import { showSuccessAlert } from './show-success-alert.js';
-import { showErrorAlert } from './show-error-alert.js';
+import { sendData } from './api.js';
 import { addImageScaling, removeImageScaling } from './image-scaling.js';
 import { addImageFiltres, removeImageFiltres } from './image-filtres.js';
-import { sendData } from './api.js';
+import { addUploadFormValidation, resetUploadFormValidation } from './upload-form-validation.js';
+import { showSuccessAlert } from './show-success-alert.js';
+import { showErrorAlert } from './show-error-alert.js';
 import { closeModal } from './show-modal.js';
-import { validateUploadForm } from './validate-upload-form.js';
 
 const formElement = document.querySelector('#upload-select-image');
-const uploadPostModal = formElement.querySelector('.img-upload__overlay');
-const uploadCencelElement = formElement.querySelector('#upload-cancel');
+const modalElement = formElement.querySelector('.img-upload__overlay');
+const cencelElement = formElement.querySelector('#upload-cancel');
 const hashtagsElement = formElement.querySelector('[name=hashtags]');
 const descriptionElement = formElement.querySelector('[name=description]');
 const submitElement = formElement.querySelector('#upload-submit');
 
+function onSubmitUploadForm(e) {
+  e.preventDefault();
+
+  const formData = new FormData(e.target);
+
+  submitElement.disabled = true;
+
+  sendData(
+    'https://26.javascript.pages.academy/kekstagram',
+    () => {
+      closeModal(modalElement);
+      showSuccessAlert();
+      submitElement.disabled = false;
+      resetUploadForm();
+    },
+    () => {
+      closeModal(modalElement);
+      showErrorAlert();
+      submitElement.disabled = false;
+    },
+    formData
+  );
+}
+
+function onFieldFocus() {
+  cencelElement.disabled = true;
+}
+
+function onFieldBlur() {
+  cencelElement.disabled = false;
+}
+
 function uploadForm() {
-  function onFieldFocus() {
-    uploadCencelElement.disabled = true;
-  }
-
-  function onFieldBlur() {
-    uploadCencelElement.disabled = false;
-  }
-
-  function onSubmitUploadPostForm(e) {
-    e.preventDefault();
-
-    const formData = new FormData(e.target);
-
-    submitElement.disabled = true;
-
-    sendData(
-      'https://26.javascript.pages.academy/kekstagram',
-      () => {
-        closeModal(uploadPostModal);
-        showSuccessAlert();
-        submitElement.disabled = false;
-        resetUploadForm();
-      },
-      () => {
-        closeModal(uploadPostModal);
-        showErrorAlert();
-        submitElement.disabled = false;
-      },
-      formData
-    );
-  }
-
   hashtagsElement.addEventListener('focus', onFieldFocus);
   hashtagsElement.addEventListener('blur', onFieldBlur);
   descriptionElement.addEventListener('focus', onFieldFocus);
   descriptionElement.addEventListener('blur', onFieldBlur);
-  formElement.addEventListener('submit', onSubmitUploadPostForm);
+  formElement.addEventListener('submit', onSubmitUploadForm);
 
   addImageScaling();
   addImageFiltres();
-  validateUploadForm();
+  addUploadFormValidation();
 }
 
 function resetUploadForm() {
   removeImageFiltres();
   removeImageScaling();
-  formElement.reset();
+  resetUploadFormValidation();
+  hashtagsElement.value = '';
+  descriptionElement.value = '';
+  formElement.removeEventListener('click', onSubmitUploadForm);
 }
 
-export { uploadForm };
+export { uploadForm, resetUploadForm };
